requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# OpenAI for embeddings
openai>=1.0.0

# Vector database options (choose one)
# chromadb>=0.4.0  # Local vector DB
# pgvector>=0.2.0  # PostgreSQL with vector support
# pinecone-client>=2.2.0  # Cloud vector DB

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Document processing
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
faiss-cpu>=1.7.0

# Optional: For better text processing
# spacy>=3.6.0
# nltk>=3.8.0

# Web framework (for API endpoints)
fastapi>=0.100.0
uvicorn>=0.23.0

# Database (if using PostgreSQL)
# psycopg2-binary>=2.9.0
# sqlalchemy>=2.0.0
